+++ source/extensions/omni.flux.stage_manager.factory/omni/flux/stage_manager/factory/items.py
@@ -15,13 +15,21 @@
 * limitations under the License.
 """
 
-from typing import Any
+from __future__ import annotations
 
-from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
+__all__ = ["StageManagerItem"]
+
+import threading
+from typing import TYPE_CHECKING, Any, Optional
+
+from omni.flux.utils.common import reset_default_attrs
+
+if TYPE_CHECKING:
+    from .plugins.tree_plugin import StageManagerTreeItem
 
 
 class StageManagerItem:
-    def __init__(self, identifier: Any, data: Any = None, children: list["StageManagerItem"] | None = None):
+    def __init__(self, identifier: Any, data: Any = None, parent: Optional["StageManagerItem"] = None):
         """
         An item that should be built by a context plugin and used by the interaction plugin and any of its children
         plugins.
@@ -29,21 +37,26 @@ class StageManagerItem:
         Args:
             identifier: An identifier for the item.
             data: Data associated with the item.
-            children: Any children items that should be under the item in the interaction plugin tree.
+            parent: The parent item
         """
         for attr, value in self.default_attr.items():
             setattr(self, attr, value)
 
         self._identifier = identifier
         self._data = data
-        self._children = children or []
+        self._parent = parent
+
+        self._is_valid = None
+        self._tree_item = None
+
+        self._lock = threading.Lock()
 
     @property
     def default_attr(self) -> dict[str, None]:
         return {
             "_identifier": None,
             "_data": None,
-            "_children": None,
+            "_parent": None,
         }
 
     @property
@@ -63,12 +76,68 @@ class StageManagerItem:
         return self._data
 
     @property
-    def children(self) -> list["StageManagerItem"]:
+    def parent(self) -> Optional["StageManagerItem"]:
         """
         Returns:
-            Any children items that should be under the item in the interaction plugin tree.
+            The parent item if one exists.
         """
-        return self._children
+        return self._parent
+
+    @property
+    def is_valid(self) -> bool:
+        """
+        Returns:
+            Whether the item is valid or not based on the active filters.
+        """
+        return self._is_valid
+
+    @is_valid.setter
+    def is_valid(self, value: bool):
+        """
+        Set the item state based on the active filters.
+
+        Updating an item will also update the parent if it exists and the item is valid.
+
+        Args:
+            value: The new value for the item.
+        """
+        with self._lock:
+            if self.is_valid is True:
+                return
+
+            self._is_valid = value
+
+            # Check if we need to update the parent while we hold the lock
+            update_parent = value is True and self.parent
+
+        if update_parent:
+            self.parent.is_valid = value
+
+    @property
+    def tree_item(self) -> StageManagerTreeItem:
+        """
+        Returns:
+            The TreeView item for the stage item.
+        """
+        return self._tree_item
+
+    @tree_item.setter
+    def tree_item(self, value: StageManagerTreeItem):
+        """
+        Set the TreeView item for the stage item.
+
+        Args:
+            value: The new TreeView item.
+        """
+        self._tree_item = value
+
+    def reset_filter_state(self):
+        """
+        Reset the filter state of the item.
+
+        Should be used before filtering the item to make sure an updated state is set.
+        """
+        self._is_valid = None
 
     def destroy(self):
-        _reset_default_attrs(self)
+        reset_default_attrs(self)