+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Repo Tools to the latest public versions
 - Refactored the Stage Manager to optimize performance and add flexibility
 - Added `IsCapture` Widget & Filter to the `AllPrims` interaction in the Stage Manager
+- REMIX-3640: Use Async Threaded Processing for Filtering of the Stage Manager items
 
 ### Fixed
 - REMIX-3401: Fixed hot-reload by allowing reuse of validators