+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.9.0]
+### Added
+- Added Timer Icon
+
+### Fixed
+- Fixed `ignore_prim_paths` for the lights interaction
+
 ## [1.8.0]
 ### Added
 - Added IsCapture widget & filter to the AllPrims interaction