+++ source/extensions/omni.flux.stage_manager.plugin.context.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.6.0]
+### Changed
+- Build a flat list of data instead of a tree to match the updated StageManagerItem data structure
+
 ## [1.5.0]
 ### Changed
 - `get_items` function now returns `StageManagerItem` objects instead of arbitrary data